gdata: Make WeakPtrFactory the last parameter
This is a reland of r149576, which was reverted as it broke
RemoteFileSystemExtensionApiTest.RemoteMountPoint. Turned out that
ui_weak_ptr_ in GDataFileSystem shouldn't be removed.

Remove WeakPtr members in favor of WeakPtrFactory::GetWeakPtr()
Along the way, copy derat's comment from ContactDatabase on this matter.

Wez on [chromium-dev] Weak ptrs for derived types?

You need to be similarly careful to have WeakPtrFactory appear after any
fields that its WeakPtrs may be used to access if you follow that pattern
instead.

BUG=none
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10843043

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@149585 4ff67af0-8c30-449e-8e8b-ad334ec8d88c