Add serial bulk reads.
As a curious shortcut, we originally implemented the serial read function to
return exactly one byte at a time. This might have been back when we didn't
yet have ArrayBuffer support, so it actually was easier to deal with a byte
rather than a buffer. Now we read the requested number of bytes or a smaller
number to avoid blocking.

BUG=136897
TEST=changed existing tests to use new API, and replaced the test byte with a test sequence.


Review URL: https://chromiumcodereview.appspot.com/10827123

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@149718 4ff67af0-8c30-449e-8e8b-ad334ec8d88c