Make the getAttention badge grey.
The grey badge fades in using the existing animation, and when the
user clicks on it, it snaps to the full-color version, without an
animation.

Also avoided granting access on right clicks, which I noticed because
of the color change.

This change has the side-effect that pageaction.show() will cause the
badge to fade in.

This doesn't yet show the grey background, as that takes
platform-specific code.

BUG=133142


Review URL: https://chromiumcodereview.appspot.com/10823143

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@149846 4ff67af0-8c30-449e-8e8b-ad334ec8d88c