contacts: Add ContactStoreFactory and FakeContactStore.
This adds a factory interface for ContactStores and a fake,
in-memory implementation of ContactStore.  Both will be
needed for testing the upcoming ContactManager class.

I'm also fixing a bug in GoogleContactStore (all contacts
were being passed to ContactDatabase in partial updates
instead of just the updated ones) and another bug in
FakeContactDatabase::SaveContacts() that was hidden by the
GoogleContactStore bug.

BUG=128805
TEST=none
TBR=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10850033

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@149924 4ff67af0-8c30-449e-8e8b-ad334ec8d88c