Use hi-resolution favicon variants if available.
This changes the favicon IPC to return all favicon variants to the browser,
and hands the to the new in-browser variant selection function.

Note that favicons that come from the history database will still
be lodpi. Reloading a site will force the favicon to be loaded from
the network. (chrome:// favicons are always loaded through the favicon
cache and hence are always lodpi)

BUG=138550
TEST=Go to apple.com, amazon.com, github.com, notice hidpi favicon

Review URL: https://chromiumcodereview.appspot.com/10828127

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@150184 4ff67af0-8c30-449e-8e8b-ad334ec8d88c