gdata: Save screenshot to /drive.
Note that this does not support the full functions of screenshots.
It merely makes possible to save the screenshot .png file to  /drive.
What are *not* yet included in this patch are:
- Show/upload the images from "Report issue" (crbug:140622).
- Taking care of name collisions. For now, when a screenshot is taken
  after a short time period of another screenshot, the new comer will
  overwrite or recorded as ".. (1).png", depending on timing.

Not quite good, but for M22, I hope it is better than nothing.

BUG=138593

Review URL: https://chromiumcodereview.appspot.com/10830179

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@150204 4ff67af0-8c30-449e-8e8b-ad334ec8d88c