Simplify devtools code on android and enable devtools for android content_shell.
This CL removes the unnecessary RemoteDebuggingcontroller java class and associated
c++ code. Also devtools_server.* is moved from content/ to chrome/ so that the
code to access profiles, version etc. can be handled here to make DevToolsServer
a self contained class in chrome/ layer handling devtools for android. The code
making use of this class is not yet ready to be upstreamed due to dependencies
and will land in future.

Also added shell_devtools_delegate_android.cc to enable devtools for content_shell
on android. This class and chrome/browser/android/devtools_server.cc need a
common place to put the IsUserAllowedToConnect method that ensures only adb is
allowed to access devtools on android (via the abstract unix socket, for security
purposes) hence that method is placed in DevToolsHttpHandler in content/

BUG=136682,136318
TEST=manual. Run android content shell, execute "adb forward tcp:9222 localabstract:content_shell_devtools_remote" and open "localhost:9222" on the host machine to use devtools.

Review URL: https://chromiumcodereview.appspot.com/10832112

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@150336 4ff67af0-8c30-449e-8e8b-ad334ec8d88c