Remove the effects of --disable-restore-session-state.
The "Better Session Restore" features have been on by default for quite a while,
and there are no known problems.

This CL removes the effects of --disable-restore-session-state but keeps the
flag and the about:flags flag in place. New experimental features are ready to be
enabled by default, and this flag will be used again to disable them.

BUG=116947


Review URL: https://chromiumcodereview.appspot.com/10855022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@150541 4ff67af0-8c30-449e-8e8b-ad334ec8d88c