Diagnostics UI: apply UI styles.
Add a navigation pane on the left (similar to the uber frame for history/extensions/settings pages) and apply the same style settings.

Note: the navigation pane is non-functional at the moment. For Diagnostics V1 (1st milestone), we will only have a single page for connectivity, which is always selected as the current page.

BUG=139442
TEST=lumpy


Review URL: https://chromiumcodereview.appspot.com/10824223

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@150793 4ff67af0-8c30-449e-8e8b-ad334ec8d88c