Do not create login items if the user deleted a previous one.
Added a new kUserRemovedLoginItem pref that is used to track if the user has
previously removed a login item for chrome. This is so we don't keep re-creating
the login item every time the user enables/installs a background app if they
have indicated that they don't want chrome to open at login.

BUG=140017
TEST=Install background app on mac, uncheck "Open at Login" checkbox in dock menu, disable and re-enable the app via Tools->Extension, make sure Open at Login checkbox does not get re-enabled.


Review URL: https://chromiumcodereview.appspot.com/10823221

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@151102 4ff67af0-8c30-449e-8e8b-ad334ec8d88c