Clean-up inline members of nested classes (chrome/)
Due to a bug, the Clang-plugin style checker failed to warn about
inline constructors, destructors, non-empty virtual methods, etc.
for nested classes.

The plugin has been fixed, and this patch is part of a clean-up of all
the code that now causes the plugin to issue errors.

BUG=139346


Review URL: https://chromiumcodereview.appspot.com/10855075

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@151135 4ff67af0-8c30-449e-8e8b-ad334ec8d88c