Fix handling of inline autocomplete text vis-a-vis Instant.
Consider the following scenario:
1. Type "g" in the omnibox.
--> The omnibox inline autocompletes "oogle.com".
2. Hit Tab.
--> You're now in keyword mode.

Here, inline_autocomplete_text_ is still "oogle.com", though the display text
is a blank string. The previous code caused a crash. Now, we check properly
that the inline text is a suffix of the display text.

Arguably, the inline text should always be a suffix of the display text, say by
clearing it in OmniboxEditModel::AcceptKeyword(): http://crbug.com/141890

R=pkasting@google.com
BUG=141878
TEST=Scenario in CL description above shouldn't crash.

Review URL: https://chromiumcodereview.appspot.com/10824261

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@151146 4ff67af0-8c30-449e-8e8b-ad334ec8d88c