Remove observer when removing delegate, in Instant.
I believe the crash (see bug link) is caused because we remove the delegate
connection, but we never stop observing the WebContents. I haven't reproduced
the crash, but I think this should fix the root cause. I've put in a bandaid as
well, which might be redundant, but good for safety. Will remove it once I can
confirm the fix works.

BUG=141875
R=sky@chromium.org
TEST=This type of crash (see bug) should go away.

Review URL: https://chromiumcodereview.appspot.com/10829284

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@151173 4ff67af0-8c30-449e-8e8b-ad334ec8d88c