sync: raise initial backoff interval from 1 second to 5 minutes.
Long term, we'd like to drop this further (i.e. 1 minute) and trust THROTTLED responses in the event of widespread backend outages.
We'd like to run some stress tests of that infrastructure (at scale / under load), but in the meantime we want to address the problem
for beta channel users / M22.  We chose 5 minutes, which is a little more aggressive than the recommended value, but we believe is
safe -- happy to discuss offline.

This patch is very low risk (stability wise). although 5 minutes is a long time to wait for users who hit this due to transient failures.
Current data suggests 500s occur around 0.02% of the time, and although backoff isn't limited to 500s it's unlikely to get hit frequently,
and the user experience doesn't break down significantly.

Effectiveness wise, We can run an experiment once it hits dev channel by returning 500s and watching for expected drops in QPS.

BUG=139733,142029

Review URL: https://chromiumcodereview.appspot.com/10826194

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@151216 4ff67af0-8c30-449e-8e8b-ad334ec8d88c