Pass parsed ChangeList to GDataFileSystem.
Get ChangeList using Drive V2 API and then converts to
DocumentFeed.  This is not optimal but change can be minimal
to make file list appear in FileBrowser UI, which makes
other development/debugging/refactoring easier.

This depends on the change in
https://chromiumcodereview.appspot.com/10837201/
for GetKind() function.

BUG=chromium:127728
TEST=specify "--enable-drive-v2-api" and open file browser (Ctrl-M), then go to drive folder and see if file listing works.

Review URL: https://chromiumcodereview.appspot.com/10837201

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@151445 4ff67af0-8c30-449e-8e8b-ad334ec8d88c