Support multiple disable reasons in ExtensionPerfs
Use a bitmask to represent all the reasons an extension is disabled.
Switch to a new key in ExtensionPrefs for the bitmask.
When we init ExtensionPrefs, check for the old key, migrate the reason over
to the new key, and delete the old key.

BUG=142762


Review URL: https://chromiumcodereview.appspot.com/10828316

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@151785 4ff67af0-8c30-449e-8e8b-ad334ec8d88c