Search state transitions not working
This change codifies and fixes the state transitions used in
the Search UI.  It documents the transitions in tabular
form in search_types.h detailing the user interaction that
results in either animated or non-animated transitions.
The pending navigation signal is removed, so all transitions now
occur on the navigation "committed" signal.

BUG=133529
TEST=Manual
R=sky@chromium.org, kuan@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10837240

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@152047 4ff67af0-8c30-449e-8e8b-ad334ec8d88c