Cocoa: Implement media gallery dialog
This is the cocoa version of the media galleries permission dialog. The dialog is implemented as an NSAlert displayed as a constrained window. The gallery checkboxes are added inside a accessory view.

The views and GTK versions are here:
http://codereview.chromium.org/10828166/
https://chromiumcodereview.appspot.com/10826129

Screenshots: http://i.imgur.com/fedm0.png

BUG=134929
TBR=mihaip@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10834242

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@152128 4ff67af0-8c30-449e-8e8b-ad334ec8d88c