Don't show the button for hidden page/browser actions or script badges.
Browser Action, Page Actions and Script Badges can be hidden, which causes us to get invalid |button| and |frame| (respectively). We'll need to figure out what to do in these cases (on all platforms) but the minimum for now is to ignore the keystroke to avoid a crash. 

BUG=140396
TEST=None
Review URL: https://chromiumcodereview.appspot.com/10828364

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@152321 4ff67af0-8c30-449e-8e8b-ad334ec8d88c