There were two leaks detected here.
The first is from a preexisting bug. GpuProcessHost::SendOnIO will leak the message if the IO thread is not present.  Note that it will not leak the message if the IO thread is present, but the GPU process is absent.

The second issue is that we would keep around one GpuData-Observer indefinitely.  I changed this to be scoped to the task manager.

On the second issue I have a concern here that there may be a race between destruction of the task manager and the GpuDataManager.  The TaskManagerModel dtor calls into GpuDataManager::RemoveObserver, but is this safe if the GpuDataManager's dtor has been called first?  Do we offer any guarantees of this?  In practice, neither singleton's dtor is called when I just run the browser, and the GpuDataManager's dtor is called second in the unit test.

BUG=143563

Review URL: https://chromiumcodereview.appspot.com/10861018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@152847 4ff67af0-8c30-449e-8e8b-ad334ec8d88c