Remove LinkDoctor from Chromium.
We want to be able to limit the requests to this URL (e.g.,
by using an API key). To do that, we need to hide the URL.

The URL has moved to an internal repository.
This change depends on the src-internal/DEPS change.

The check box is still in the Options page because that check
box also controls the captive portal feature.

BUG=143327


Review URL: https://chromiumcodereview.appspot.com/10832434

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@152908 4ff67af0-8c30-449e-8e8b-ad334ec8d88c