Don't record browser window display UMA if first run dialog is shown
Missed this in http://codereview.chromium.org/10662046/ . Don't record the Startup.BrowserWindowDisplay UMA if the first run UI is displayed.

This is the only other place I could find in the code that was displaying a custom dialog on startup on Win/Mac.  The previous CL added this call to the ShowMessageBox() call so all other dialogs should be covered.

BUG=None
TEST=None


Review URL: https://chromiumcodereview.appspot.com/10887006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@153770 4ff67af0-8c30-449e-8e8b-ad334ec8d88c