Add new (possibly shorter) expiry time for prerenders that lose their launchers.
When a user navigates away from a page containing a prerender, this code adds a new shorter timeout. Currently it's 30s (same as the default timeout), but this will have an effect on the 5minTTL group, as they will also have a 30s navigate away timeout.

R=dominich@chromium.org,mmenke@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/10870070

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@153788 4ff67af0-8c30-449e-8e8b-ad334ec8d88c