drive: Simplify DriveFileSystem::GetFileByResourceId()
This is a reland of crrev.com/153776, which was reverted due to crashes
in tests, which caused by misuse of scoped_ptr<>::Pass() again.

The number of callers of GetEntryByResourceIdAsync() in DriveFileSystem
is reduced from 2 to 1.

Along the way, make |get_file_callback| parameter mandatory for GetFileByPath()
GetFileByResourceId(), and related helper functions.

BUG=143518
TEST="make available offline" works as before, which is implemented on top of this function

Review URL: https://chromiumcodereview.appspot.com/10887017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@153928 4ff67af0-8c30-449e-8e8b-ad334ec8d88c