Fix some zoom bubble DCHECK() failures
When a user clicks on the zoom icon on Windows, if there was already a bubble
showing that bubble is hidden and replaced.  By the time that bubble's
WindowClosed() event happens, the static pointer (zoom_bubble_) is no longer the
same as this, so the DCHECK() was failing.

Additionally, when zooming with your mouse over the zoom bubble, a timer was
stopped but being asked to be Reset() (which also DCHECK() fails).

BUG=None (yet)
R=khorimoto@chromium.org,pkasting@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10896033

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@154059 4ff67af0-8c30-449e-8e8b-ad334ec8d88c