Remove searchTerms from the Instant URL.
We would always replace searchTerms with a blank string when loading the
Instant URL, so it never served any useful purpose, except to confuse
maintainers about whether or not the Instant template was a valid one to
consider for search terms extraction.

Also make a minor change to the instantEnabledParam so that it appends a "&"
instead of prepending it, thus being more consistent with other params.

BUG=139176
R=dominich@chromium.org,pkasting@chromium.org,sky@chromium.org
TEST=none; no change in functionality.

Review URL: https://chromiumcodereview.appspot.com/10823417

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@154255 4ff67af0-8c30-449e-8e8b-ad334ec8d88c