contacts: Add ContactManagerInterface::GetWeakPtr().
The ContactManager singleton is destroyed in
PostMainMessageLoopRun(), which appears to run before the
(not-yet-checked-in) ContactProvider class is destroyed in
response to the user profile being destroyed.  This change
lets ContactProvider hold a weak pointer to ContactManager
so it can avoid accessing it during shutdown.

I'm also renaming a couple of references to "partial
updates" to "incremental updates" for consistency.

BUG=128805,141877
TEST=none


Review URL: https://chromiumcodereview.appspot.com/10896031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@154327 4ff67af0-8c30-449e-8e8b-ad334ec8d88c