Reliably install default apps into new installations.
r139107 changed the check to see if the profile that we're running it was
created by the current Chrome version. This does not appear to be true for the
default profile created during the first run (it doesn't have an explicit
profile.created_by_version preference, and the default value is 1.0.0), so
bring back the IsChromeFirstRun check (the profile version check is still kept
for additional profiles that are created).

BUG=145351
R=asargent@chromium.org
TEST=do a clean install of Chrome, the default profile should have the Gmail, YouTube and Search apps. Same for additional profiles that are created.

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=154450

Review URL: https://chromiumcodereview.appspot.com/10916041

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@154532 4ff67af0-8c30-449e-8e8b-ad334ec8d88c