Constrained Windows Cocoa: Part 1 Custom controls
This CL adds new controls that match the look and feel of the new constrained windows.

CWButton is a new button that looks similar to the Chrome WebUI button. It has a shadow and glows on hover.

CWWindow is a new window type that is borderless and has a rounded shape.

CWAlert is a replacement for NSAlert that uses CWButton and CWWindow. It also uses a custom layout to match the new look and feel.

Screenshots: http://i.imgur.com/80MGA.png
             http://i.imgur.com/3JrII.png

BUG=140537, 140539, 140562
TEST=Verified that all unit tests passed.

Review URL: https://chromiumcodereview.appspot.com/10828412

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@154644 4ff67af0-8c30-449e-8e8b-ad334ec8d88c