Fix preedit handling on IME switching.
This is hack for Korean IME which required to commit preedit when it is disabled or switched to other IME.
The root of the problem is that, in ChromeOS, if the next engine is xkb layout, SetGlobalEngine is not called.
Due to this logic, ibus-daemon can't know the ime switching.
As the result, IBusPreeditFocusMode does not work.

BUG=chrome-os-partner:12798
TEST=Manually done on lumpy

Review URL: https://chromiumcodereview.appspot.com/10917061

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@154802 4ff67af0-8c30-449e-8e8b-ad334ec8d88c