Centralize logic around Instant modes.
The goal of this change is to have callers go through InstantController to check
if it's enabled (in any of a variety of modes), as opposed to checking
prefs::kInstantEnabled directly. This is because we plan to add field trials,
command-line forced modes and default modes that may cause Instant to be in a
given mode regardless of this pref. At the moment however, there's no change in
functionality.

BUG=none
R=pkasting@chromium.org,sky@chromium.org
TEST=none; no change in functionality, as the SUGGEST mode is not actually used at the moment.

Review URL: https://chromiumcodereview.appspot.com/10879043

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@154809 4ff67af0-8c30-449e-8e8b-ad334ec8d88c