webui: Request favicons at all required display scale factors
Note: This doesn't help for webui pages that set the favicon
to a lodpi png using <link rel=icon>. The options page does
this to switch the favicon for different views.

For this reaosn, remove an uneeded link rel=icon from the
downloads page.

BUG=145228

Review URL: https://chromiumcodereview.appspot.com/10908096

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@154991 4ff67af0-8c30-449e-8e8b-ad334ec8d88c