uber: Explicitly add 2x favicons
pngs can only contain a single resolution. The alternative would
be to add a chrome://theme-ico/ source that returns ico files
(which can contain several bitmaps) and use that for favicons.
But there's no ico-writing code in the browser yet, so let's go
with this simpler solution for now.

BUG=145228
TEST=Go to preferences, history, extensions, help using chrome
on a machine with a retina display. Notice the hidpi favicons.

Review URL: https://chromiumcodereview.appspot.com/10928002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@155030 4ff67af0-8c30-449e-8e8b-ad334ec8d88c