Add an API to retrieve the keys that services should use for Google APIs.
This tightens up the DEPS for chrome/common/net/gaia so that it depends
on none of Chrome, and the same is true for the new 
chrome/common/net/google_apis.  The intent is to make it reusable by 
components outside of chrome/, such as remoting/, but it would be nice
to follow up with a move to a better location for sharing.

This change switches Cloud Print, Remoting, and any user of GaiaClient
to using keys via the new API.  It does not however change any of the
key values or the way they are provided (although it does introduce
means of overriding them).

Follow-up work will make it easy to specify the keys via include.gypi
or chromium.gyp_env, and add baking-in of keys for some build types.

BUG=145584
TEST=Any service that requires log-in could be affected by this, e.g. Cloud Print, Chromoting, and sync. The change should have no impact on these services; in particular you should not need to log in again after you update to a build with this change if you were previously logged in.

Review URL: https://chromiumcodereview.appspot.com/10889044

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@155100 4ff67af0-8c30-449e-8e8b-ad334ec8d88c