Network connectivity - packet pacing and FEC tests for packet loss.
After finishing the regular experiment for 21 packets loss test, we
start one of the following two new tests.

1) Pace the packets. The delay between the packets is calculated from
the total time to receive packets in the first experiment (total time
is the time from when we recieved the first packet to the time it has
taken to recieve the last packet). If the total time is 20 secs, then
we send each packet with a delay of 1 sec.

2) Non-paced test. Will send packets as soon as possible.

Deleted TCPStatsClient class and thus no TCP connectivity
tests. Merged UDPStatsClient into NetworkStats class.

R=jar
TEST=unit tests

Review URL: https://chromiumcodereview.appspot.com/10913037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@155111 4ff67af0-8c30-449e-8e8b-ad334ec8d88c