mac: Make the throbber 25% more efficient
I added a local change that made all tabs always show throbbers.

I noticed that -getRectsBeingDrawn:count: for a TabView would
return two rects while a throbber is active: (18, 5, 16, 16),
(the throbber subview rect; since the throbber is not an
opaque view, the tab view below it needs to be repainted),
and a mysterious 1-px wide 16-px high rect pretty far right
in the tab.

What happened was that the tab to the left of the tab with the
throbber overlapped the throbber by one pixel, and since TabView
isn't opaque either, the tab to the left needed repainting too.

The dirty rect passed to drawRect: is the union of all invalidated
rects, and the union of the throbber rect with the 1px wide rect
to the far right is almost the whole tab.

Reduce tab overlap by 1 pixel, to fix this. This reduces cpu usage
with 8 throbbing tabs (viewing about:blank) from 20% cpu to 15% cpu
on my retina MBP while plugged in, and from 30% cpu to 20% cpu while
not plugged in.

BUG=19268


Review URL: https://chromiumcodereview.appspot.com/10916147

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@155236 4ff67af0-8c30-449e-8e8b-ad334ec8d88c