[Sync] Generalize StringOrdinal to handle ordinal_in_parent field
Rename StringOrdinal to Ordinal, move it to sync/, and templatize it.
Make StringOrdinal be an instantiation of Ordinal that matches its
previous behavior.

Create NodeOrdinal, which is another instantiation of Ordinal for
the ordinal_in_parent field in SyncEntity.

Rework Ordinal to handle arbitrary byte ranges and to simplify the
interpolation code a bit.

Generalize StringOrdinal unit tests for Ordinal.

Update users of StringOrdinal to prepend syncer::.

BUG=145412
TBR=estade@chromium.org,jianli@chromium.org,brettw@chromium.org, 

Review URL: https://chromiumcodereview.appspot.com/10920017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@155368 4ff67af0-8c30-449e-8e8b-ad334ec8d88c