autocomplete: Add ContactProvider.
This adds an AutocompleteProvider implementation that
searches through the user's contacts.  It will eventually be
used to provide results in the Chrome OS app list.

ContactProvider built only on Chrome OS at the moment (as
that's also the only platform where the rest of the contacts
code is built) and is not instantiated outside of tests.
I'm also punting on ranking the results; all matches get a
relevance of 1.

BUG=141877
TEST=none
TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10907066

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@156224 4ff67af0-8c30-449e-8e8b-ad334ec8d88c