Reland: Take 2: Force python test server output to be unbuffered, so it doesn't mix with gtest output
In browser tests that use a local python server, the python output in
the test logs sometimes overlaps with gtest output, resulting in gtest
falsely detecting passing tests as incomplete. This is a result of
python's default use of buffered output, which gets written to the log
file out of order.

This patch forces the python process for local test servers to use
unbuffered mode via the -u switch. This way, by the time gtest is ready
to log a passing test, all testserver output is already written to the
log file.

Update: The win xp bots were adversely affected by the -u switch, so we now
apply it only for mac and linux, which is where the original problem was found.

BUG=147368
TEST=See sync integration test output when it is redirected to a log file, and make sure there are no false negatives.

Originally committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=156361

Review URL: https://codereview.chromium.org/10907162

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@156389 4ff67af0-8c30-449e-8e8b-ad334ec8d88c