Add the power interface extension API. This CL adds the power interface extension API. The API is experimental and is accessible via the chrome.experimental.power.x interface. The API proposal is available on the API proposals page at http://dev.chromium.org/developers/design-documents/extensions/proposed-changes/apis-under-development
The implementation of the API uses user preferences to store the keep_awake requests for a user. This makes the API specific to a user and persistent between logins. Currently we're leaving the preference unsync'able to keep them local to the machine the extension is installed on.

R=satorux@chromium.org,derat@chromium.org,aa@chromium.org
BUG=146939
TEST=Ran an extension to call requestKeepAwake and cancelKeepAwake to verify that the system power management changes accordingly.


Review URL: https://chromiumcodereview.appspot.com/10919270

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@157010 4ff67af0-8c30-449e-8e8b-ad334ec8d88c