Use a single flag for enabling experimental WebKit features.
Merge multiple command line webkit features into a single flag.  This
has the benefit that when a new feature is added, we can add it to this
flag rather than requiring users to toggle another flag.  This is
similar to what extensions does for experimental APIs.

BUG=148503


Review URL: https://chromiumcodereview.appspot.com/10917269

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@157014 4ff67af0-8c30-449e-8e8b-ad334ec8d88c