Launch wallpaper manager/picker in a container and hide app icon from app list
According to PM, we want to launch the wallpaper manager in a 640x440 window.
And we also want to fit 5 thumbnails in a row.

BUG=149250
TEST=
launch wallpaper manager from context menu on a chromeos device. Verify it the width and
height is 640x440. (note: the window frame height should also included in the 440)

Review URL: https://codereview.chromium.org/10907235

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@157114 4ff67af0-8c30-449e-8e8b-ad334ec8d88c