Add guards to metric values; erase bad events/metrics from db
Improves the database by not inserting blatantly erroneous data, and by deleting
any events or metrics which cannot be properly retrieved from the database.
Also make DB take a metric for AddMetric() - this makes it so that we have a
more accurate timestamp since we record the time before the thread-jumping.

BUG=130212


Review URL: https://chromiumcodereview.appspot.com/10907121

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@157168 4ff67af0-8c30-449e-8e8b-ad334ec8d88c