Bluetooth: Defer observer methods for properties until have address
The AdapterPoweredChanged and AdapterDiscoveringChanged observer
methods could be called while the adapter had no address, and was
claimed not to be present, potentially confusing the UI. Defer
until the address is obtained, and the adapter announced as present.

Add lots of property-related unit tests.

BUG=chromium-os:34498
TEST=unit_tests
Review URL: https://codereview.chromium.org/10910309

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@157229 4ff67af0-8c30-449e-8e8b-ad334ec8d88c