Fixed freezes while downloading to Google Drive.
This patch fixes the issue, when downloading files larger than 1 MB which are already in browsers cache, directly to Google Drive. In most cases it is just big html files.

This fix introduces DriveUploader::OnUploaderReady callback, which will be called when uploader gets ready. Then we can force updating the upload from the DriveDownloadObserver.

The problem existed, because formerly we used DriveDownloadObserver::UpdateUpload method to update the upload process. However in some cases (see above) all of the DriveDownloadObserver::UpdateUpload invocations did not update the upload because the uploader was never ready. So, as a result the upload process was never completed.

BUG=145831

Review URL: https://chromiumcodereview.appspot.com/10911180

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@157302 4ff67af0-8c30-449e-8e8b-ad334ec8d88c