Start the process of making Favicon into a Browser Component.
This means, add the intended DEPS rules for a Browser Component
(depend on nothing within chrome/browser outside the component except
chrome/browser/api) and add a temporary owner for pure refactoring
changes so that such changes can be done more quickly. Larger changes
(that affect functionality or interfaces in non-refactoring ways) will
still be sent to a proper OWNER.

With the new DEPS, a presubmit check will warn developers that add new
temporarily-allowed (the files designated with a "!" rule in the DEPS
file).

BUG=146851


Review URL: https://chromiumcodereview.appspot.com/10928053

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@157370 4ff67af0-8c30-449e-8e8b-ad334ec8d88c