Updates the Chrome OS file browser to support opening files with Web Intent handlers. Adds a basic test.
This is added to the list of 'normal' file handlers in the same way that Drive tasks are currently added. (although I feel like the whole flow needs to be changed a bit, since we don't store defaults for anything but the original file browser tasks).

BUG=138664

Review URL: https://chromiumcodereview.appspot.com/10834383

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@157693 4ff67af0-8c30-449e-8e8b-ad334ec8d88c