Move the NOTIFICATION_APP_TERMINATING notification from content to chrome, since it's fired by chrome.
I removed the code watching it in the workers devtools code. It's not clear if this code is needed or not, so removing it for now. If we get crash reports or memory leaks on bots, then we'll have a repro that we can use to fix this.
Review URL: https://codereview.chromium.org/10968064

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@158566 4ff67af0-8c30-449e-8e8b-ad334ec8d88c