Revert 158707 - Enable by default: 1. threaded compositing on windows (except XP) 2. force compositing mode on mac
On linux we switch to a 1/3 1/3 fcm and thread experiments .
This patch also gives priority to the command line switches so that both threaded
and force compositing mode can be turned off via the about:flags page or command line.

In the near future, we will use the Finch framework to tweak the experiment percentages.

Note: This patch has a dependency on:
https://bugs.webkit.org/show_bug.cgi?id=96871

It also disables ExtensionApiTest.CaptureVisibleTabJPeg as it triggers flakiness on it. chrome.tabs.captureVisibleTab is inherently racy as there is no way to guarantee that the tab has been painted at least once when that call is made. Note how the other variants of CaptureVisibleTab tests have already been disabled for similar reasons.

BUG=143809,128387,146861

Review URL: https://chromiumcodereview.appspot.com/10914247

Reverted due to failures on Mac bots:
http://build.chromium.org/p/chromium.mac/builders/Mac10.6%20Tests%20%281%29/builds/29446/steps/interactive_ui_tests/logs/AutoResize
http://build.chromium.org/p/chromium.mac/builders/Mac10.7%20Tests%20%281%29/builds/2143/steps/interactive_ui_tests/logs/TightAutosizeAroundSingleLine
http://build.chromium.org/p/chromium.mac/builders/Mac%2010.6%20Tests%20%28dbg%29%283%29/builds/26358/steps/interactive_ui_tests/logs/TightAutosizeAroundSingleLine
http://build.chromium.org/p/chromium.mac/builders/Mac%2010.7%20Tests%20%28dbg%29%283%29/builds/1804/steps/interactive_ui_tests/logs/TightAutosizeAroundSingleLine

TBR=vangelis@chromium.org
Review URL: https://codereview.chromium.org/10991032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@158711 4ff67af0-8c30-449e-8e8b-ad334ec8d88c