Implement popup bubbles for the controlled setting indicator
This CL adds support for popup bubbles to the settings UI and makes the options.ControlledSettingIndicator show such a bubble when it is clicked. The options.Bubble class inherits from cr.ui.BubbleBase, allowing all style and most of the code to be shared between the different bubbles in Chrome's Web UI.

The CL also fixes several long-standing bugs in the Chrome OS classes options.ManagedNetworkIndicator and options.NetworkList so that the bubbles these pop up behave consistently with the rest.

TEST=manual, including UI in rtl languages
BUG=104955

Review URL: https://chromiumcodereview.appspot.com/10907148

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@158783 4ff67af0-8c30-449e-8e8b-ad334ec8d88c