Give platform apps control over launcher right-click context menu.
This moves most of the extension context menu related code from RenderViewContextMenu
to a separate class, which is then used in two other places as well.

This time without a bug that caused the 'tab' parameter in the onClicked event to
no longer get valid values (as well as an extension to a test so this problem is
not just caught in debug builds but also in tests run in release builds).
See http://codereview.chromium.org/10918103/ for the previous go at this patch.

BUG=143222

Review URL: https://chromiumcodereview.appspot.com/10979036

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@159312 4ff67af0-8c30-449e-8e8b-ad334ec8d88c