Histogram - add check for external Histogram data with mismatching range.
This adds an extra step of validation for Histogram data coming from an external
data source, similar to the check in http://crrev.com/154887.

If a histogram already exists for the specified name but  the range of the new
data does not match the existing histogram, then discard the new data to prevent
CHECK failure in Histogram::FactoryGet().

BUG=153328
TEST=local build works, passes tests on CQ

Review URL: https://codereview.chromium.org/11032005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@159538 4ff67af0-8c30-449e-8e8b-ad334ec8d88c