Fail incoming requests in PepperFlashSettingsManager after an error.
This is the second of two CLs to address the issue mentioned below. It is in itself enough in itself to fix the issue, but given that the issue actually discovered two bugs, there is an other CL to fix the other bug too. See issue for details.

BUG=144874

Review URL: https://chromiumcodereview.appspot.com/10986059

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@159569 4ff67af0-8c30-449e-8e8b-ad334ec8d88c