Cache the channel ID
To prevent hitting rate limits on how fast we can access the server to get an
auth token to get a channel Id, we cache the channel Id when we get one.
Note that a channel Id is per profile, so we store them by profile login name.


BUG=149733


Review URL: https://chromiumcodereview.appspot.com/10951022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@159576 4ff67af0-8c30-449e-8e8b-ad334ec8d88c