Added MediaGalleriesPreferencesTest.UpdateGalleryType test.
Whenever a gallery is added, its type is set either to "AutoDetected"
or "UserAdded". When the gallery is removed, auto detected galleries type is set to
"BlackListed" and user added galleries are deleted. When the gallery is added again, its type is updated from
"BlackListed" to "UserAdded".

BUG=none
TEST=none


Review URL: https://chromiumcodereview.appspot.com/10986075

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@159619 4ff67af0-8c30-449e-8e8b-ad334ec8d88c