Rename InvalidationState to BootstrapData
We have a lot of things called 'state' in sync/, so we're renaming them to
better reflect their actual purpose. The 'state' in this CL is used for
bootstrapping a connection to the cache invalidation server, so we don't need
to go through the process of re-registering each ID we're interested in each
time we reconnect.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/11050005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@159845 4ff67af0-8c30-449e-8e8b-ad334ec8d88c