Fixed the browsingData API so that it now actually performs deletion of server-bound certificates.
Modified the implementation so that now it parses the arguments to the remove() general-purpose function correctly, that is, it now looks for an option named "serverBoundCertificates" instead of "serverBoundCerts", which is now in line with the JSON API definition.

As per our discussion with mkwst, I removed implementation for the removeServerBoundCerts() function (which was never defined in the JSON API), and modified the removeCookies() short-cut function to remove server-bound certificates as well (and updated the documentation accordingly). Rationale: it makes a lot more sense to remove server-bound certs along with cookies, and the remove() function still provides a way to remove them separately if need be. 

Also fixed a subtle race condition that prevented multiple concurrent browsing data removal sessions from being prevented.

BUG=152948

Review URL: https://chromiumcodereview.appspot.com/10987091

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@159886 4ff67af0-8c30-449e-8e8b-ad334ec8d88c