Finish implementation of controlled setting indicator
This CL completes the implementation of a controlled setting indicator
class for Chrome's settings UI. This class can be used to show an icon
next to each setting whose value is enforced or recommended by policy or
an extension. Clicking on it or selecting the indicator by keyboard brings
up an informative bubble.

The CL ensures that indicators already used in the Chrome OS network
settings remain functional. No further indicators are added to the
settings UI - this will be done in follow-up CLs.

BUG=104955

TEST=Chrome OS VM with ONC-configured VPN; only here is a controlled
TEST=setting indicator currently being used (see crbug.com/137470)


Review URL: https://chromiumcodereview.appspot.com/11066015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@160351 4ff67af0-8c30-449e-8e8b-ad334ec8d88c