Move timestamp de-duping logic from HistoryBackend to NavigationController
Encapsulate the de-duping logic in a new class TimeSmoother.  Improve
the logic to avoid dupes completely unless the clock jumps backwards.

Note that moving the de-duping logic to NavigationController means that timestamps
are de-duped per-tab.  But this is okay given the problem we're trying to solve (which I added a comment about).

Add function to override timestamp generation for testing.

BUG=128449


Review URL: https://chromiumcodereview.appspot.com/11043028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@160578 4ff67af0-8c30-449e-8e8b-ad334ec8d88c