Do not modify homepage prefs on first run
This CL removes a hack that allowed existing and new users to have
different default homepages. Since the NTP is the default homepage for all
users anyway these days, the hack is no longer needed. Removing it has the
benefit that it allows the homepage to be recommended through policy as it
is no longer explicitly set in the profile.

BUG=154603
TEST=Start with fresh profile, observe no homepage set in Preferences


Review URL: https://chromiumcodereview.appspot.com/11088007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@160689 4ff67af0-8c30-449e-8e8b-ad334ec8d88c