[Sync] Protect against tabs without valid navigations.
If the UI doesn't update, or a race occurs and a tab loses its navigations
while the restore task is pending, it's possible we'll attempt to restore
a tab that has no navigations, leading to a crash.

Similarly, if a foreign sessions is partially deleted, we'll lose the tab
information, but still create default-initialized SessionTabs within the
SessionWindow. "Open all" will attempt to restore those tabs too, which
should not trigger crashes.

BUG=154129


Review URL: https://chromiumcodereview.appspot.com/11038072

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@160752 4ff67af0-8c30-449e-8e8b-ad334ec8d88c