USB extension API: retrieve data from isochronous packet descriptors.
Currently the USB extension API only considers the actual_length member of the libusb_transfer structure to determine the size of the received data. However, in the case of isochronous transfers this value might be zero yet data is available inside the different isochronous packet descriptors of the struct. This patch fixes the issue.

BUG=154576


Review URL: https://chromiumcodereview.appspot.com/11066056

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@161011 4ff67af0-8c30-449e-8e8b-ad334ec8d88c