drive: Split IsGDataAvailable() into two functions
IsGDataAvailable() was a misnomer. It should be named like IsDriveEnabled()
as this function was designed to check if Drive was enabled or not.

However, this function was also used in google_contact_store.cc. There,
we should only check if the given profile can be used for authentication
for Google APIs, hence AuthService::CanAuthenticate() is introduced in
this patch.

BUG=153962
TEST=none


Review URL: https://chromiumcodereview.appspot.com/11085037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@161278 4ff67af0-8c30-449e-8e8b-ad334ec8d88c