Fix Cloud Print sign in flow.
Somewhere along the way, the sequence of events that get triggered when
completing the sign in flow changed and was no longer being caught in a way
that would cause the sign in dialog to close properly. Update the code that
watches for sign in completion to the new way of doing things.

BUG=147321
TEST=Open Chrome with a fresh profile (no user logged in), visit a page, bring
up print preview (not the native print dialog), hit the 'Change' button to
bring up the printer list, click on the sign in link at the bottom, go through
the flow. Before this CL, sign in would get stuck and the cloud print jobs page
would open in the background tab. After, sign in should go away and the printer
list will refresh with cloud printers from the signed in account.


Review URL: https://chromiumcodereview.appspot.com/11106002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@161461 4ff67af0-8c30-449e-8e8b-ad334ec8d88c