Makes gfx::Screen an instance, rather than a collection of static methods.
This is in support of supporting separate Screen implementations on Aura for desktop and metro on Windows.

Some callsites are not yet correct, and noted with a reference to the http://crbug.com/133312. As-is those sites will behave the same as before this patch, but may not be correct once desktop/metro can run simultaneously.


BUG=133312


Review URL: https://chromiumcodereview.appspot.com/11030017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@161644 4ff67af0-8c30-449e-8e8b-ad334ec8d88c