Mac: Fix inline extension install prompt
Installing inline extensions stopped working on Mac.

In r161306 I switched the extension install prompt to be a constrained window.

This requires a window to attach the install prompt onto. The inline install code path was not setting a parent window causing us not to show the prompt.

Fix was to pass a window.

BUG=155732
TEST=Navigate to http://chrome.angrybirds.com/ and choose "Install to play offline". Verified that the install prompt was shown.


Review URL: https://chromiumcodereview.appspot.com/11145010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@161961 4ff67af0-8c30-449e-8e8b-ad334ec8d88c