Fixed: on windows, fixed-size app windows would lose their shadows.
Some code checked the resizability of the window and used it to key a change in
window style. For app windows, we always want a very particular window style,
so we check for the remove_standard_frame Widget::InitParams option and ignore
the resize flag.

Also, change the shell window NonClientHitTest code to not return HTTOPLEFT and
friends when the window is non-resizable.

R=ben@chromium.org
BUG=155219


Review URL: https://chromiumcodereview.appspot.com/11154007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@162387 4ff67af0-8c30-449e-8e8b-ad334ec8d88c