Shift passage of FileStream in downloads system to be by scoped_ptr<>.
http://codereview.chromium.org/10912173/ constructs the DownloadFile, and
thus the BaseFile, on the UI thread and then passes it to the FILE thread.
DownloadFile / BaseFile may be constructed with a FileStream to which to 
write the download.  The FileStream cannot be passed by linked_ptr<> in this
case, as that is not thread safe. 

BUG=123998
R=benjhayden@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11028131

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@162411 4ff67af0-8c30-449e-8e8b-ad334ec8d88c