Clears the list of apps to restore when the browser starts normally.
In the case of a browser restart, we restart the apps that were running when the browser exited. This change makes it so that when chrome is starting normally (ie: not restarting) we clear the list of running apps, so that subsequent chrome restarts don't prompt the restart of apps that are no longer running.


BUG=156243


Review URL: https://chromiumcodereview.appspot.com/11187016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@162683 4ff67af0-8c30-449e-8e8b-ad334ec8d88c