Fix an issue where odd-sized windows would have their bounds rounded up on Mac.
Specifying that an app window should have a size of 359x359 caused the created
window to be 360x360 if an x and y coordinate was not specified, because the x
and y coordinates were being calculated as half the width of the screen in
floating point (i.e. they would end up being fractional).

BUG=151985
R=dharcourt@chromium.org,avi@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11185071

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@162944 4ff67af0-8c30-449e-8e8b-ad334ec8d88c