DevTools: merge is_docked and dock_side into single dock_side state.
As of today, we manage devtools window state in a weird way: it can be either docked or undocked, and while docked it can attach to bottom or to the right. These two levels of setting makes it hard to manage code, we'd like devtools window to have enum of equal states: bottom, right, undocked (so far). This change makes it happen.

Review URL: https://chromiumcodereview.appspot.com/11189060


git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@162982 4ff67af0-8c30-449e-8e8b-ad334ec8d88c