Instant Extended flags cleanup and change in default values
This CL changes the default values of the "Show White NTP" and "Show Logo"
settings to be true and true.  I've changed the names of these prefs
so that old values are forgotton and all users are forced to these new
defaults.

This also cleans up the naming of the various prefs associated with the
instant-extended UI variants.

BUG=none
TEST=Launch the old way, launch the new way, observe new defaults kick in.
R=sreeram@chromium.org, estade@chromium.org, pkasting@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/11227013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@163286 4ff67af0-8c30-449e-8e8b-ad334ec8d88c