Upstream Blocked Popup Content Settings changes
Since blocked popups rely on an info bar instead of a persistent button, we have to handle
the case where the info bar might have been dismissed and make sure to show it again.  So
every time we block a popup we should notify the WebContents so the front end has a chance
to handle the event and make sure an info bar is showing.

BUG=http://crbug.com/155685


Review URL: https://chromiumcodereview.appspot.com/11238008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@163358 4ff67af0-8c30-449e-8e8b-ad334ec8d88c