Renamed files in preparation for adding calculator controller tests
To support testing the controller logic of the calculator default app,
that logic has to move out of view.js. It will be moved into
background.js, which will then become a regular controller. Also, when
support is added for testing the controller logic, the tests will become
general calculator tests rather than just model tests. In accordance
with this, the following files were renamed:

    app/background.js -> app/controller.js
    tests/model_tests.html -> tests/tests.html
    tests/model_tests.js -> tests/tests.js
    tests/model_test_utilities.js -> tests/utilities.hml

That renaming is done in this commit so the actual code changes which
will be made in a future commit will not be obscured by the name
changes. The code changes made for this commit were the minimal ones
necessary for the file name changes.

BUG=156695


Review URL: https://chromiumcodereview.appspot.com/11272012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@164004 4ff67af0-8c30-449e-8e8b-ad334ec8d88c