Make Chrome OS OOBE update check screen announce accessible progress.
When this screen appears it now speaks "Checking for updates".
If it takes a while, you hear 'period' as the ellipses appear and
disappear, so at least you know it's doing something; we should
improve this later, but this is a good first step.

BUG=127062


Review URL: https://chromiumcodereview.appspot.com/11270017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@164048 4ff67af0-8c30-449e-8e8b-ad334ec8d88c