cros: Disable login UI while loading user profile.
User profile is loaded asynchronously while signing in. When the loading is
in progress, accessibility, IME, volume etc could not be changed because they
access extension system and it should not be accessed before profile finishes
loading. Disable UI so that user would not be able to change them.

Fundamental issue is tracked in http://crbug.com/148878.

BUG=157885
TEST=User should not be able to change accessibility/IME/volume/brightness
after sign-in starts and before it finishes. And no crash should happen when
user attempts those.

R=nkostylev@chromium.org

Review URL: https://codereview.chromium.org/11269062

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@164401 4ff67af0-8c30-449e-8e8b-ad334ec8d88c