Add "dump to file" button to about:sync.
In the Data tab, there's a "dump to text" button. This adds a
"dump to file" button which triggers a download of a bunch of
debugging data suitable for attaching to a bug.

The format is fake CSV. There's a header with date/time plus
useragent, followed by field labels, and then the data.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11232067

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@165073 4ff67af0-8c30-449e-8e8b-ad334ec8d88c