Set root resource ID upon full feed update.
We used to set the resource ID at DriveResourceMetadata's
constructor as for WAPI the resource ID for root is
a fixed string, but for Drive API it is not fixed and
was set when About resource is obtained.

This caused a trouble for maintaining the code and
this CL is trying to consolidate setting root resource
ID at one place.

Although for Drive API 'root' can be used for a convenient
alias for queries, Drive API usually returns a reference
to root directory (e.g. reference to parent folder)
as its real root directory resource ID rather than 'root'
alias, therefore we don't use it here.

BUG=152230, 152288, 157314
TEST=compiles, pass all existing tests.


Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=164876

Review URL: https://chromiumcodereview.appspot.com/11227020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@165219 4ff67af0-8c30-449e-8e8b-ad334ec8d88c