Reenable the BookmarkBarViewTest8.DNDBackToOriginatingMenu test on AURA.
While instantiating the Shell class for the duration of the tests, we pass in a ShellDelegate
which is implemented by the TestShellDelegate class. This should ensure that the tests run without
assuming that the screen is locked. 

The code to remove registered EventClients has been removed.

BUG=158564
R=ben
Review URL: https://codereview.chromium.org/11347051

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@165239 4ff67af0-8c30-449e-8e8b-ad334ec8d88c