Enable thumbnails to be stored after tab is closed on Aura.
This is a preparation for taking thumbnails when tab is closed.
Currently this works only on Aura. On Mac, the asynchronous callback is just
canceled when tab is closed. On the other platforms, the callback runs
synchronously and this change does not affect the behavior.

BUG=125400
TEST=Manually check that the thumbnail is stored after the tab is closed.


Review URL: https://chromiumcodereview.appspot.com/11106006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@165401 4ff67af0-8c30-449e-8e8b-ad334ec8d88c