Ensure that keyboard input works in Chrome AURA on Windows 8 in metro. cpu: please review everything, sky: Please review the changes to ui and chrome\browser.
We need to send character events to the RenderViewHost class in addition to the KeyDown/Keyup events
for this to work.

To achieve this we register the character received event on the Windows 8 CoreWindow class and send the
character info via a new IPC message MetroViewerHostMsg_Character.

The chrome side handling for this message is performed in the MockInputMethod::DispatchFabricatedKeyEvent
function where we dispatch the message to the focused TextInputClient. 

While working on this I also noticed that we were not sending the event flags like shift/alt/Ctrl, etc to
Chrome from the Metro code. Added support for that. This ensures that we are able to select text in the edit
fields on the page and the omnibox, etc.

The above changes will most likely not work with IME's. Will look into that in a subsequent change.

BUG=158637
R=cpu,sky
Review URL: https://codereview.chromium.org/11348053

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@165496 4ff67af0-8c30-449e-8e8b-ad334ec8d88c