Fixed setting focus in settings' popups.
This patch fixes a bug, when opening popups in Settings page and not updating focus. The problem was especially significant when ChromeVox was enabled.

With this patch, focus is set to the first focusable control in a popup. This is now default behavior, which can be overriden though, to focus any another control.

TEST=1. Enter Settings, 2. Go to show 'advanced settings' using tab and enter keys, 3. Enter Content Settings using enter key.
BUG=158884


Review URL: https://chromiumcodereview.appspot.com/11363030

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@165637 4ff67af0-8c30-449e-8e8b-ad334ec8d88c