[Autofill] Disable the browser process popup UI for non-Android platforms on M24.
This UI is not baked enough on Windows or Linux to be enabled in M24.  However, it is the only Autofill UI for Android, and should remain on for that platform.

This CL is meant to be committed, merged to M24, then reverted on trunk.

BUG=158371

Review URL: https://codereview.chromium.org/11364058

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@165809 4ff67af0-8c30-449e-8e8b-ad334ec8d88c