File Manager: Change the mount status even when the mount is not requested.
If the mount is requested from C++ side (instead of UI side),  the mount status will not be changed at mountGData() because mountCompleted event is occurred at first and mountGData() is never called. In such a case, the mount status is changed at onMountCompleted() instead of mountGData().

BUG=158022
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/11312002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@165899 4ff67af0-8c30-449e-8e8b-ad334ec8d88c