drive: FeedToEntryProtoMap shouldn't have a side-effect.
While converting a feed to EntryProtoMap, the function was overwriting
upload URL of the root entry. This patch splits the conversion part and
the URL update part. Since the function is applied also for non-root
feeds, the side effect has corrupted the resource metadata.

BUG=159087

Review URL: https://chromiumcodereview.appspot.com/11312098

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@166172 4ff67af0-8c30-449e-8e8b-ad334ec8d88c