Update PluginFinder#FindPluginWithIdentifier return value.
Updating FindPluginWithIdentifier to correctly return false in case PluginInstaller is requested to be output (by setting the output parameter |installer| to non-NULL value) and it was not found. This can only happen in case an installer for an unknown plugin (we have no metadata for it) is requested.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/11367111

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@166331 4ff67af0-8c30-449e-8e8b-ad334ec8d88c