First couple of steps of improving the extension/app update process.
For Apps and Extensions with a lazy background page, this will delay installing
an update until the background page gets unloaded (or next chrome restart).
Also adds a chrome.runtime.onUpdateAvailable event which is fired when an update is
available but isn't being installed immediately because the app is not idle.
It doesn't change anything yet for Extensions with a persistent background page
(or extensions without a background page at all). Those still get updated as soon
as an update is found.

BUG=37971
BUG=143260
TBR=nick@chromium.org 

Review URL: https://chromiumcodereview.appspot.com/11293002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@166358 4ff67af0-8c30-449e-8e8b-ad334ec8d88c