CPM Data Fetching Refactor
This allows for batch data fetching (i.e. getting data for metrics a, b, and c
in one call) rather than single-type data fetching.

This will increase performance, and is necessary for future improvements.

Additionally, replace all cases of AggregationStrategy (and related) with AggregationMethod, since before both these terms were being used.

Note: While this is a large change, it is not quite as large as Rietveld would have you believe. Many changes (particularly in performance_monitor_handler.cc) result in only a change in indentation, which counts as a full-line change. Much of the code is actually identical to what it was.

Review URL: https://chromiumcodereview.appspot.com/11363049

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@166815 4ff67af0-8c30-449e-8e8b-ad334ec8d88c