Allowing the extension to run in incognito should not affect whether the browser action icon is shown.
We used to make sure that the browser action was always visible if the user enabled the extension. We did this by modifying the visibility flag in the ExtensionService (in EnableExtension). This is too far down in the code, since things like 'allow in incognito' reload the extension, causing EnableExtension to be called. Instead, we should do this only as a result of the user flipping the Enabled flag in the Extensions page handler.

BUG=155177
TEST=This repros all the time with the repro steps listed in the bug, so take a look at the bug first. Then hide a browser action, and disable and enable through the Extension settings page and the browser action should become visible again.
Review URL: https://codereview.chromium.org/11360122

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@166910 4ff67af0-8c30-449e-8e8b-ad334ec8d88c