Fix Omnibox UI Test to Properly Defined Behavior
In the interactive UI test suite, OmniboxViewTest's DeleteItem test is incorrect.  It depends on the history URL provider's willingness under a certain condition to suggest an unvisited intranet site.  HistoryURL provider should never do so.

I changed this test to use an unvisited path on a known host.  These are allowed to be suggested.

Incidentally, the test currently happens to work due to a complicated interaction between prevent_inlining, the fact that the unvisited intranet site happens to have a history entry, and that therefore the on-disk URL database classifies the site as visited because it's been typed, even though its visit count is 0.  I'm sure this test wasn't intended to test this complicated confluence of factors, so I altered the URLs for this test so we it doesn't hit this edge case.

BUG=

Review URL: https://chromiumcodereview.appspot.com/11275201

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@166923 4ff67af0-8c30-449e-8e8b-ad334ec8d88c