[Autofill] Rename GetInfo and SetInfo to GetRawInfo and SetRawInfo
This is part 1 of a multi-CL change to always read the application locale from
an appropriate thread in the Autofill code.  The next CL will rename
GetCanonicalizedInfo and SetCanonicalizedInfo to GetInfo and SetInfo, and add a
|locale| parameter to these methods.  Subsequent CLs will migrate all Autofill
code that currently access the locale to always access it from the UI thread.

BUG=100845
TEST=none (pure refactoring)

TBR=dhollowa@chromium.org, akalin@chromium.org, bauerb@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11360055

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@167072 4ff67af0-8c30-449e-8e8b-ad334ec8d88c