Combining the SpellCheckHost and the SpellCheckProfile into the SpellcheckService.
There is still a lot of additional cleanup I would like to do (e.g., removing all the functions which simply pass along to the particular dictionaries) after this one is in. This one is primarily a lot of shuffling of code around and fixing naming to be more logical.

Note: I will delete the vestigial files in a separate CL.


BUG=154923
TEST=test have been renamed: spellcheck_service_unittest.cc, spellcheck_service_browsertest.cc

TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11364190

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@167210 4ff67af0-8c30-449e-8e8b-ad334ec8d88c