Settings App WebUI description string substitution, "re"localization. This change introduces the mechanism for subtituting strings in the Settings App WebUI to allow for the different wording when the settings webui elements are used for the settings app, and sample strings.
This introduces OptionsPageUIHandler::RegisterConstStrings, a helper
function to remove some of the boilerplate around substituting product
names into loadTimeData.

The approach for substituting strings for the settings app is by populating
a dictionary within loadTimeData which is used as a source of replacements
in JS (by adding a JS function LoadTimeData.overrideValues). This allows
the same strings.js to be used, and cached, on the browser side.

BUG=130210
TEST=chrome://settings unchanged, chrome://settings-frame/options_settings_app.html
should show new wording for, e.g., Users (on an empty profile): "You are currently
the only App Launcher user."

Review URL: https://chromiumcodereview.appspot.com/11339035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@167886 4ff67af0-8c30-449e-8e8b-ad334ec8d88c