Making it easier to hook a ContentViewCore to the SurfaceView.
This CL changes the way we deal with WebLayers, so that a ContentViewCore does
not need native code anymore to hook with the surface view.
The ContentViewCore now keeps track of the WebLayers as they are
attached/detached. The ContentViewRenderView is used to render a specific
ContentView by retrieving the ContentViewCore WebLlayer and adding it to the
compositor.

BUG=
TEST=ContentShell and ChromiumTestShell should still work.


Review URL: https://chromiumcodereview.appspot.com/11293264

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@168166 4ff67af0-8c30-449e-8e8b-ad334ec8d88c