Allow the --install-from-webstore flag to work when chrome is already running.
This requires the flag to be handled later in the startup sequence, at a point that is executed for cold startups, and also for chrome handing control over to an existing, running process.

This change also means chrome does not return a status code signifying the success or failure of the operation.

BUG=160607


Review URL: https://chromiumcodereview.appspot.com/11413004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@168169 4ff67af0-8c30-449e-8e8b-ad334ec8d88c