Prevent redownloading of already downloaded crx files.
Now that extension updates might not be installed immediately after downloading
and unpacking it was possible that ExtensionUpdater/Downloader would download
the same update again in a new update check. This attempts to fix that by taking
the version of a pending update into account when determining what updates to
download.

BUG=161353


Review URL: https://chromiumcodereview.appspot.com/11316029

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@168424 4ff67af0-8c30-449e-8e8b-ad334ec8d88c