Enable users to alter their custom dictionary from an overlay on chrome://settings/editDictionary. This overlay can also be accessed through:
Settings
|_Show advanced settings...
..|_Language and spell-checker settings...
....|_Custom spelling dictionary

When a user types an unknown word into a textarea, the browser underlines this word as a misspelling. The user can right-click on this word and select "Add to Dictionary". If the user made a mistake, the user needs a way to remove a word from their custom dictionary. This CL provides this functionality.

Automated tests are in https://codereview.chromium.org/11280013/ and https://codereview.chromium.org/11308076/.

BUG=18238

Review URL: https://chromiumcodereview.appspot.com/11362063

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@168760 4ff67af0-8c30-449e-8e8b-ad334ec8d88c