Optimize getting preference values within PrefService class.
The PrefService class spends a lot of time getting Preference objects when all it wants is the value of the named preference.
GetPreferenceValue() is an optimized version of GetPreference()->GetValue(), and is now used whereever a preference value is needed within
the PrefService class.

This reduces the Android startup time by about 30ms.

BUG=161252


Review URL: https://chromiumcodereview.appspot.com/11420007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@168797 4ff67af0-8c30-449e-8e8b-ad334ec8d88c