Let chromeos::User decide whether the current session can be locked
We have several different methods that try to determine whether the
current session can be locked. This redundancy is difficult to maintain
and has drifted out of sync in the past. The CL consolidates all checks
and makes the chromeos::User class the autoritative source for whether
the current session can be locked or not.

BUG=152928


Review URL: https://chromiumcodereview.appspot.com/11412100

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169031 4ff67af0-8c30-449e-8e8b-ad334ec8d88c