Remove self-destruct message box (causing hang in Metro).
Launch self-destruct setup out of job (so that it doesn't die when killing the chrome.exe which created it...).

And be more drastic about killing Chrome processes on force-uninstall (pure kill, no wait).


BUG=158632,159995
TEST=Install system-level over user-level (with user-level set to run in Metro mode); next run should self-destruct user-level Chrome (silently); subsequent launches should launch system-level Chrome.
This should also fix users currently in this bad state (upon updating the user-level Chrome which will then be able to self-destruct properly).

This might also solve bug 159995, will need to test with official build (it definitely solves the part where when both Chromes are present, user-level is activated, but hangs on the self-destruct message (which can't be displayed in Metro) -- the stronger kill process in this CL might also help with the user-level binaries left behind issue).


Review URL: https://chromiumcodereview.appspot.com/11417036

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169130 4ff67af0-8c30-449e-8e8b-ad334ec8d88c