Use PostTaskAndReply and a WeakPtr to handle the case of ImageLoader being deleted before loading an image has finished.
Turns LoadImagesOnBlockingPool into a non-member method, and also removes some dead code
from there. ReplyBack is kept as a member since even if it doesn't access any ImageLoader
state at the moment (there is no state), if in the future we might want to add caching,
ReplyBack will be the point to store data in the cache.

BUG=141673


Review URL: https://chromiumcodereview.appspot.com/11421032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169155 4ff67af0-8c30-449e-8e8b-ad334ec8d88c