Remove PrerenderManager::DoShutdown(), have PrerenderUnitTest do proper two phase shutdown.
The PrerenderManager::DoShutdown() method was confusing, since it was really just another way of accessing the ProfiledKeyService::Shutdown() method. So, let's remove that, and while we're at it, having the unit test do a two phase shutdown.

This patch is preparation for the prerender events/observer change that's coming.

R=mmenke@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/11411157

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169450 4ff67af0-8c30-449e-8e8b-ad334ec8d88c