google_apis: Remove a hard-coded URL hack from HttpRequestParser
"http://localhost/" was hard-coded in HttpRequestParser, but this was wrong
as the server runs at a different host name and a port.

Along the way, add a test for HttpServer::RegisterDefaultResponse, which was
missing. Also add the request URL to warning messages for 404 requests.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/11280156

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169576 4ff67af0-8c30-449e-8e8b-ad334ec8d88c