Have PrerenderManager::AddPrerender take ownership of PrerenderContents earlier.
By putting the PrerenderData into active_prerenders_ immediately after creating it, we can be sure that when we enter PrerenderData::MoveEntryToPendingDelete that the underlying prerender is in active_prerenders_, and so there always will be an owning PrerenderData, making the lifetime of PrerenderContents much closer in line to PrerenderData. This simplifies MoveEntryToPendingDelete by reducing the number of paths through this tricky function.

The one exception to this ownership rule is in unit_tests. By making MoveEntryToPendingDelete virtual, like other methods in PrerenderManager for testing, we catch those and put that special case logic inside of the unit test. The browser test doesn't need to override the main body, so it gets test coverage there.

This change is based on https://codereview.chromium.org/11411157/ , and has to commit after it.

R=mmenke@chromium.org
BUG=None

Review URL: https://chromiumcodereview.appspot.com/11299165

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169686 4ff67af0-8c30-449e-8e8b-ad334ec8d88c