In which PrerenderManager::PrerenderData unfriends PrerenderManager.
The friendship wasn't needed with a few more methods on PrerenderData, and that helps make the code a lot more readable.

While we're here, make the pending delete list a ScopedVector of PrerenderData, since we've now gotten rid of the cases where the pending delete entries were not in the active map.

R=mmenke@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11421076

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169737 4ff67af0-8c30-449e-8e8b-ad334ec8d88c