Make PanelExtensionBrowserTest.ClosePanelBeforeIconLoadingCompleted pass on win aura.
The issue is that on win aura the root window is destructed asynchronously. Painting through the compositor is also asynchronous. So the window starts getting closed, and before it's destroyed, painting happens. There's null-dereference at PanelFrameView::GetFaviconForTabIconView() because window_ was being set to NULL. Since the window is still valid, only set it to NULL when the WidgetObserver callback indicates it's being destroyed. Instead of using window_ being NULL as a signal that closing has already happened, add another bool to keep track of that.

BUG=154081
Review URL: https://codereview.chromium.org/11428038

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169832 4ff67af0-8c30-449e-8e8b-ad334ec8d88c