Have ExtensionService notify ThemeService directly for themes
Save the theme ID if the service has not yet been created, as creating
the service causes a startup regression.

This is so that the ThemeService correctly picks up the new theme
even if it hasn't been created yet.

Also clean up the various getters/setters a bit.

BUG=161203


Review URL: https://chromiumcodereview.appspot.com/11348238

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@169875 4ff67af0-8c30-449e-8e8b-ad334ec8d88c