Revert 170030 - Add OAuth2TokenService to fetch oauth2 access tokens for given scopes using the oauth2 refresh token maintained by TokenService.
This cl adds a service that fetches oauth2 access tokens for given scopes using the oauth2 refersh token maintained by TokenService. The fetched tokens are kept in a cache and when a token fetching request is received, the cached token is returned if there is a valid one cached. The token fetching request can be started from any thread and it can be canceled.

BUG=NONE


Review URL: https://chromiumcodereview.appspot.com/11362136

TBR=chenyu@chromium.org
Review URL: https://codereview.chromium.org/11308251

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170035 4ff67af0-8c30-449e-8e8b-ad334ec8d88c