bookmarks: Pass BrowserContext to ToggleWhenVisible() function.
We are switching away from Profile, and content::BrowserContext is preferred
over it.

This patch was moves it to ui/bookmarks, because it's only used by c/b/ui,
and renames it to ToggleBookmarkBarWhenVisible() so it's more clear what is being
toggled for the consumers.

BUG=144783
R=sky@chromium.org

Review URL: https://codereview.chromium.org/11411228

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170167 4ff67af0-8c30-449e-8e8b-ad334ec8d88c