Reenable GeolocationBrowserTest.ErrorOnPermissionDenied
This test was marked crashy, but unfortunately the stack trace
in the bug report is insufficient to diagnose the error. The test
is not crashing on my local machine, so let's reenable this in
the hope that it is either not crashing anymore, or that we can
get a better stack trace.

TEST=GeolocationBrowserTest.ErrorOnPermissionDenied
BUG=70585


Review URL: https://chromiumcodereview.appspot.com/11348252

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170282 4ff67af0-8c30-449e-8e8b-ad334ec8d88c