Eliminates all ash dependencies from WindowModalityController.
This is in preparation for moving it to corewm so this code can be shared with desktop-aura.

This required moving GetActivatableWindow to ActivationClient (instead of ash-specific ActivationController) and consolidating the modal parent property.

For a subsequent CL: moving the test's dependencies (ChildModalWindow, CaptureTrackingView) from ash/test to views/test.

http://crbug.com/158115
R=rouslan@chromium.org
Review URL: https://codereview.chromium.org/11418224

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170313 4ff67af0-8c30-449e-8e8b-ad334ec8d88c