ash: Make WebKit caret blink speed match Views.
This updates NativeTextfieldView's insert cursor/caret to be
shown for 500 ms and hidden for 500 ms when blinking.
Previously, it was 800/500, but these numbers were
apparently chosen arbitrarily.  I'm also making the 500 ms
interval get passed to WebKit so that carets in web content
will blink at the same speed as the caret in the omnibox.

BUG=162261
TEST=caret blinks at same speed in omnibox and on google.com


Review URL: https://chromiumcodereview.appspot.com/11308237

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170353 4ff67af0-8c30-449e-8e8b-ad334ec8d88c