Immersive mode reveals the tabstrip/omnibox on top of web content
ImmersiveModeController reparents the tabstrip and toolbar to a new RevealView
during the slide-down reveal of the top views. This allows the view hierarchy
in the non-immersive-mode case to remain as it is, reducing the likelihood this
change will break Windows.

BrowserViewLayout stops caching pointers to tabstrip and toolbar as it can
access them directly from BrowserView.

BUG=162310


Review URL: https://chromiumcodereview.appspot.com/11411250

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170385 4ff67af0-8c30-449e-8e8b-ad334ec8d88c