Add additional profile shortcut manager tests.
Adds a test for CreateProfileShortcutFlags() and another test for when
a second profile is created without shortcut, which verifies the first
profile's icon got renamed and badged (via icon).

Also, refactors RenameProfile() to a helper function and makes
ValidateProfileShortcutAtPath() check the icon that was set on the shortcut.

BUG=none
TEST=this adds some

Review URL: https://chromiumcodereview.appspot.com/11280240

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170408 4ff67af0-8c30-449e-8e8b-ad334ec8d88c