Fix use of FindTabbedBrowser when handling windows.create calls.
In practice, this particular call to FindTabbedBrowser will only happen for Ash-in-metro, so using the default host desktop type (as was done in http://crrev.com/168064) is guaranteed to be the wrong choice. Forcing HOST_DESKTOP_TYPE_ASH would be correct for this specific case, but using the currently active desktop (which will always be Ash) will also work and is more consistent with other places where a browser needs to be found.

BUG=129187
TBR=aa@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11412151

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170412 4ff67af0-8c30-449e-8e8b-ad334ec8d88c