Removes the |event| parameter from ActivationDelegate::ShouldActivate and makes the method const.
It is mostly unused, except for one place in RenderWidgetHostViewAura, where depending on the event type etc it will have a consequence. I think it's bad for a method with a const-sounding name like "ShouldActivate" to have side-effects.

Instead, I made the RWHVA store the data that would indicate a pointer activate and check that state during OnActivated(). It'd be a little cleaner if OnActivated had some way of knowing what the currently processed input event was, hence the TODO.

http://crbug.com/162100
R=sadrul@chromium.org
Review URL: https://codereview.chromium.org/11419251

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170477 4ff67af0-8c30-449e-8e8b-ad334ec8d88c