Omnibox: Create Field Trial for HQP to Ignore Mid-Word Matches
Adds code to HistoryQuick provider to make it optionally only consider terms when they match at the beginning of a word boundary in the URL or titles.  The current behavior allows terms to match anywhere regardless of word boundaries.

Then, creates a field trial to control this behavior.  Enables the trial for 25% of users.

BUG=161911


Review URL: https://chromiumcodereview.appspot.com/11421139

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170784 4ff67af0-8c30-449e-8e8b-ad334ec8d88c