Fix showing constrained window for uninitialized tabs
If a tab is added without being shown then the associated tab view for the web content will not be created yet.

When a constrained window is added to that tab the window never gets shown.

To fix this I've changed ConstrainedWindowMac2 to reshow the window once the web contents becomes visible.

BUG=163778
TEST=Reproduced bug 163778 and verified that my patch fixes it.
Verified that ConstrainedWindowMacTest.ShowInUninitializedTab fails without my patch and passes with it.


Review URL: https://chromiumcodereview.appspot.com/11419277

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170795 4ff67af0-8c30-449e-8e8b-ad334ec8d88c