chromeos: Remove GViewRequestInterceptor.
This removes the code for sending the user to GView for PDF
and .ppt files.  The --enable-gview flag has been passed
unconditionally on Chrome OS for a long time, but we ignore
it in official builds.  For Chromium OS, I believe that
https://chrome.google.com/webstore/detail/docs-pdfpowerpoint-viewer/nnbmlagghjjcbdhgmkedmbmedengocbn
provides similar functionality.

BUG=161786
TBR=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11411071

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@170875 4ff67af0-8c30-449e-8e8b-ad334ec8d88c