Reland "Make wallpaper picker manifest and thumbnails available when offline."
Relands this CL: https://codereview.chromium.org/11348215/ It was reverted because of WallpaperPicker test fail.

I misused a bool to compare with -1 in the original CL. And in order to fix it,
https://codereview.chromium.org/11419315/ is landed. But the logic in 11419315
is not correct. So it cause the WallpaperPicker test failed on TOT. See detailed
test failure here: https://codereview.chromium.org/11280300

This CL fixed the logic and modified function comment to make it more clear.

Note that patchset 1 reverted the following reverts:
https://codereview.chromium.org/11280300
https://codereview.chromium.org/11434113


TBR=flackr, miket, jhawkins, thakis
BUG=158668

Review URL: https://chromiumcodereview.appspot.com/11316339

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171168 4ff67af0-8c30-449e-8e8b-ad334ec8d88c